refactor: split src/plan-runs/coordinator.test.ts into per-phase test files (warren-6bf6)
