+++ .mulch/expertise/runs.jsonl
@@ -161,3 +161,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T17:44:46.684Z","evidence":{"seeds":"warren-6fcc","commit":"fbe60a2d0c8174b5a5345cb0055caaae32de233b"},"content":"detectRuntimeTerminal (src/runs/stream.ts) is exported (warren-6fcc / pl-5516 step 2) so its branch table can be unit-tested directly in src/runs/stream.detect.test.ts without dragging the full bridgeRunStream fixture (db, repos, broker, pool). Bridge-level behavior (terminal breaks the loop, trailing events dropped) stays covered by warren-2687 in src/runs/stream.test.ts.","id":"mx-ee3fc0"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T19:44:23.364Z","evidence":{"commit":"ff57022e871290593acc75f8cd864ba4de113b64"},"relates_to":["mx-3df5c5"],"content":"Handlers monolith split phase 2 (warren-a2b4 / pl-9088 step 2): src/server/handlers/index.ts → runs.ts (789 lines) + plan-runs.ts (484 lines). Pattern matches phase 1 (mx-3df5c5): ROUTE_TABLE stays in handlers/index.ts as the canonical wire surface for gen:docs / gen:openapi; shared helpers (readJsonBody[OrEmpty], requireString, optionalString, requireParam, parseBoolean, parseNonNegativeInt, assertPlotIdDispatchable, defaultSpawn) are exported from index.ts and cross-imported. Streaming plumbing (bridgeAbort/asNdjsonStream/eventToNdjson) lives in runs.ts and is re-exported for plan-runs.ts since both NDJSON tails share it — keeping it adjacent to streamRunEventsHandler (the larger consumer) rather than hoisting to index.ts. Tests: relocated handlers.plan-runs.test.ts → handlers/plan-runs.test.ts and removed its biome noExcessiveLinesPerFunction override by splitting the monolithic POST /plan-runs describe into three siblings (core + Plot integration (warren-b89f) + plot_id validation (warren-bae5)) sharing a setupPlanRunFixture() factory. Helpers (poolFor, depsFor, makeBurrowClient) are duplicated between runs.test.ts and the residual handlers.test.ts because both files retain describes that use them — splitting helpers into a shared module would have leaked test-only plumbing across multiple files.","id":"mx-99ad0d"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T20:25:54.740Z","evidence":{"seeds":"warren-ea0c","commit":"ed47012f666066f4b3c04ddcac74b9054a6fbfd0"},"content":"reap.ts monolith split (warren-ea0c / pl-9088 step 4): src/runs/reap.ts (2114 lines) + reap.test.ts (2792 lines) → src/runs/reap/ subdirectory with per-sub-step modules. Pattern mirrors handlers split (mx-3df5c5/mx-99ad0d/mx-929bbc): index.ts is the re-export shim + module-level doc-comment; types.ts pulls the public interfaces into their own file so they can be imported without dragging the orchestrator's transitive imports; run.ts holds the reapRun orchestrator after extracting three helpers (runPreviewLaunch + runPreviewAnnotate in preview.ts, runPrOpen in pr-open.ts, dispatchAutoPlanRuns in auto-plan-run.ts) to fit under the 500-line threshold; pure helpers (mergeMulchFile, mergePlotEventsFile, mergePlotJsonFile) re-export through index.ts so external consumers (src/runs/index.ts) only change the import path suffix. Tests split alongside: each .test.ts sits next to its module and references shared fixtures via reap/test-helpers.ts (fakeFs, fakeExec, fakeBurrowClient, fakeOpenPr, setup, makePool, makeBurrow, Ctx + re-exported bun-side types). The four complex functions still over Biome's noExcessiveCognitiveComplexity ceiling (mergeMulch, mergePlot, mirrorSeeds, mirrorPlans, reapRun) carry over to the override block under their new module paths — refactoring those algorithms was explicitly out of scope for the mechanical split.","id":"mx-cdd0ca"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T04:05:22.896Z","evidence":{"seeds":"warren-6bf6","commit":"56a30a2ba61ed1f0f090eb03b9576e3384402b50"},"content":"Split src/plan-runs/coordinator.test.ts (804 lines) into coordinator.test-helpers.ts and four focused per-phase test files (coordinator.dispatch.test.ts, coordinator.gating.test.ts, coordinator.resume.test.ts, coordinator.completion.test.ts) under 500 lines to remove its file-size budget","id":"mx-bfddb2"}
