+++ scripts/file-size-budgets.json
@@ -15,7 +15,6 @@
 		"src/db/repos/runs.ts": 610,
 		"src/diagnostics/checks.test.ts": 653,
 		"src/diagnostics/checks.ts": 616,
-		"src/plan-runs/coordinator.test.ts": 804,
 		"src/plan-runs/coordinator.ts": 587,
 		"src/plots/aggregate.ts": 506,
 		"src/projects/manage.test.ts": 763,
