+++ src/plan-runs/coordinator.completion.test.ts
@@ -1,117 +1,9 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { openDatabase, type WarrenDb } from "../db/client.ts";
-import { createRepos, type Repos } from "../db/repos/index.ts";
-import type { PlanRunRow } from "../db/schema.ts";
-import { agents } from "../db/schema.ts";
-import {
-	advancePlanRun,
-	type CoordinatorEmitFn,
-	type CoordinatorShowSeedFn,
-	type CoordinatorSpawnFn,
-	type CoordinatorTransitionPlotFn,
-} from "./coordinator.ts";
+import { type Harness, NOW, neverPoll, setup } from "./coordinator.test-helpers.ts";
+import { advancePlanRun, type CoordinatorTransitionPlotFn } from "./coordinator.ts";
 import type { AutoTransitionResult } from "./plot-transition.ts";
-import type { PrMergeChecker } from "./pr-merge.ts";
 
-const NOW = new Date("2026-05-17T00:00:00.000Z");
-
-interface CapturedEvent {
-	runId: string;
-	kind: string;
-	payload: Record<string, unknown>;
-}
-
-interface Harness {
-	db: WarrenDb;
-	repos: Repos;
-	projectId: string;
-	planRun: PlanRunRow;
-	events: CapturedEvent[];
-	emit: CoordinatorEmitFn;
-	showSeedStub: (status: "open" | "closed") => CoordinatorShowSeedFn;
-	spawnStub: (newRunId: () => string) => CoordinatorSpawnFn;
-	makeRun: (seedId: string) => Promise<string>;
-}
-
-async function setup(): Promise<Harness> {
-	const db = await openDatabase({ path: ":memory:" });
-	db.drizzle
-		.insert(agents)
-		.values({
-			name: "claude-code",
-			renderedJson: { sections: {} },
-			registeredAt: "2026-05-10T00:00:00.000Z",
-			lastRefreshed: "2026-05-10T00:00:00.000Z",
-		})
-		.run();
-	const repos = createRepos(db);
-	const project = await repos.projects.create({
-		gitUrl: "https://github.com/x/y.git",
-		localPath: "/data/projects/x/y",
-		defaultBranch: "main",
-	});
-	const { planRun } = await repos.planRuns.create({
-		planId: "pl-acc",
-		projectId: project.id,
-		agentName: "claude-code",
-		children: [
-			{ seq: 1, seedId: "warren-a" },
-			{ seq: 2, seedId: "warren-b" },
-		],
-		now: NOW,
-	});
-	const events: CapturedEvent[] = [];
-	const emit: CoordinatorEmitFn = async (runId, kind, payload) => {
-		events.push({ runId, kind, payload });
-	};
-	const showSeedStub = (status: "open" | "closed"): CoordinatorShowSeedFn => {
-		return async (_projectId, seedId) => ({ id: seedId, status });
-	};
-	const spawnStub = (newRunId: () => string): CoordinatorSpawnFn => {
-		return async ({ child, prompt }) => {
-			const run = await repos.runs.create({
-				agentName: "claude-code",
-				projectId: project.id,
-				prompt,
-				renderedAgentJson: { sections: {} },
-				trigger: "plan-run",
-				seedId: child.seedId,
-				now: NOW,
-			});
-			void newRunId;
-			return { runId: run.id };
-		};
-	};
-	const makeRun = async (seedId: string): Promise<string> => {
-		const run = await repos.runs.create({
-			agentName: "claude-code",
-			projectId: project.id,
-			prompt: `work on sd ${seedId}`,
-			renderedAgentJson: { sections: {} },
-			trigger: "plan-run",
-			seedId,
-			now: NOW,
-		});
-		return run.id;
-	};
-	return {
-		db,
-		repos,
-		projectId: project.id,
-		planRun,
-		events,
-		emit,
-		showSeedStub,
-		spawnStub,
-		makeRun,
-	};
-}
-
-const neverPoll: PrMergeChecker = async () => {
-	throw new Error("checkPrMerged should not be called in this branch");
-};
-
-describe("advancePlanRun — state machine", () => {
+describe("advancePlanRun — completion phase", () => {
 	let h: Harness;
 
 	beforeEach(async () => {
@@ -120,51 +12,6 @@ describe("advancePlanRun — state machine", () => {
 
 	afterEach(async () => {
 		await h.db.close();
-	});
-
-	test("queued → running, dispatches first child", async () => {
-		const result = await advancePlanRun({
-			planRun: h.planRun,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "run_x"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-		const reloaded = await h.repos.planRuns.require(h.planRun.id);
-		expect(reloaded.state).toBe("running");
-		expect(reloaded.startedAt).toBe(NOW.toISOString());
-		const children = await h.repos.planRuns.listChildren(h.planRun.id);
-		const first = children.find((c) => c.seq === 1);
-		expect(first?.state).toBe("dispatched");
-		expect(first?.runId).not.toBeNull();
-		expect(h.events.map((e) => e.kind)).toContain("plan_run.dispatched");
-	});
-
-	test("non-terminal child run → waiting_for_run; running run syncs child.state", async () => {
-		await h.repos.planRuns.transitionTo(h.planRun.id, "running", { startedAt: NOW.toISOString() });
-		const runId = await h.makeRun("warren-a");
-		await h.repos.runs.markRunning(runId, NOW);
-		await h.repos.planRuns.updateChild({
-			planRunId: h.planRun.id,
-			seq: 1,
-			patch: { runId, state: "dispatched", startedAt: NOW.toISOString() },
-		});
-		const planRun = await h.repos.planRuns.require(h.planRun.id);
-		const result = await advancePlanRun({
-			planRun,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "unused"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("waiting_for_run");
-		const children = await h.repos.planRuns.listChildren(h.planRun.id);
-		expect(children.find((c) => c.seq === 1)?.state).toBe("running");
 	});
 
 	test("succeeded child with prUrl → pr_open then waiting_for_merge on open poll", async () => {
@@ -349,32 +196,6 @@ describe("advancePlanRun — state machine", () => {
 		expect(h.events.some((e) => e.kind === "plan_run.merged")).toBe(true);
 	});
 
-	test("resume semantics: closed seed flips to skipped without dispatch", async () => {
-		await h.repos.planRuns.transitionTo(h.planRun.id, "running", { startedAt: NOW.toISOString() });
-		const planRun = await h.repos.planRuns.require(h.planRun.id);
-		let calls = 0;
-		const showSeed: CoordinatorShowSeedFn = async (_p, seedId) => {
-			calls += 1;
-			// First call (seq=1) reports closed; second (seq=2) reports open.
-			if (seedId === "warren-a") return { id: seedId, status: "closed" };
-			return { id: seedId, status: "open" };
-		};
-		const result = await advancePlanRun({
-			planRun,
-			repos: h.repos,
-			showSeed,
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "unused"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-		expect(calls).toBe(2);
-		const children = await h.repos.planRuns.listChildren(h.planRun.id);
-		expect(children.find((c) => c.seq === 1)?.state).toBe("skipped");
-		expect(children.find((c) => c.seq === 2)?.state).toBe("dispatched");
-	});
-
 	test("plan_succeeded: every child terminal, no pending left", async () => {
 		await h.repos.planRuns.transitionTo(h.planRun.id, "running", { startedAt: NOW.toISOString() });
 		const runId = await h.makeRun("warren-a");
@@ -409,44 +230,6 @@ describe("advancePlanRun — state machine", () => {
 		expect(reloaded.state).toBe("succeeded");
 		expect(reloaded.endedAt).toBe(NOW.toISOString());
 		expect(h.events.some((e) => e.kind === "plan_run.succeeded")).toBe(true);
-	});
-
-	test("spawn receives the PlanRun.plotId so per-child PLOT_ID injection lights up", async () => {
-		// Seed a Plot-bound PlanRun directly so the coordinator's spawn
-		// closure sees a non-null plotId on the row passed through.
-		const { planRun: plotBound } = await h.repos.planRuns.create({
-			planId: "pl-plot",
-			projectId: h.projectId,
-			agentName: "claude-code",
-			plotId: "plot_acc",
-			children: [{ seq: 1, seedId: "warren-p" }],
-			now: NOW,
-		});
-		const captured: { plotId: string | null }[] = [];
-		const spawn: CoordinatorSpawnFn = async ({ planRun, child, prompt }) => {
-			captured.push({ plotId: planRun.plotId });
-			const run = await h.repos.runs.create({
-				agentName: "claude-code",
-				projectId: h.projectId,
-				prompt,
-				renderedAgentJson: { sections: {} },
-				trigger: "plan-run",
-				seedId: child.seedId,
-				now: NOW,
-			});
-			return { runId: run.id };
-		};
-		const result = await advancePlanRun({
-			planRun: plotBound,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn,
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-		expect(captured).toEqual([{ plotId: "plot_acc" }]);
 	});
 
 	test("plan_succeeded with plotId + transitionPlot → emits plan_run.plot_auto_done", async () => {
@@ -617,188 +400,5 @@ describe("advancePlanRun — state machine", () => {
 		});
 		expect(result.kind).toBe("plan_succeeded");
 		expect(called).toBe(false);
-	});
-
-	test("dispatch failure → plan_failed with dispatch_failed:<message>", async () => {
-		const failingSpawn: CoordinatorSpawnFn = async () => {
-			throw new Error("burrow unreachable");
-		};
-		const result = await advancePlanRun({
-			planRun: h.planRun,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: failingSpawn,
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("plan_failed");
-		if (result.kind === "plan_failed") {
-			expect(result.reason).toBe("dispatch_failed:burrow unreachable");
-			expect(result.failedSeq).toBe(1);
-		}
-		const reloaded = await h.repos.planRuns.require(h.planRun.id);
-		expect(reloaded.state).toBe("failed");
-		expect(reloaded.failureReason).toBe("dispatch_failed:burrow unreachable");
-		const children = await h.repos.planRuns.listChildren(h.planRun.id);
-		expect(children.find((c) => c.seq === 1)?.state).toBe("failed");
-	});
-});
-
-describe("advancePlanRun — parentRunId gate (warren-d9a2)", () => {
-	let h: Harness;
-
-	beforeEach(async () => {
-		h = await setup();
-	});
-
-	afterEach(async () => {
-		await h.db.close();
-	});
-
-	async function createPlanRunWithParent(parentRunId: string): Promise<PlanRunRow> {
-		const { planRun } = await h.repos.planRuns.create({
-			planId: "pl-parent-gate",
-			projectId: h.projectId,
-			agentName: "claude-code",
-			parentRunId,
-			children: [
-				{ seq: 1, seedId: "warren-c" },
-				{ seq: 2, seedId: "warren-d" },
-			],
-			now: NOW,
-		});
-		return planRun;
-	}
-
-	test("no parentRunId → existing behavior, dispatches immediately", async () => {
-		const result = await advancePlanRun({
-			planRun: h.planRun,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "run_x"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-	});
-
-	test("parentRunId set, parent PR merged → gate passes, dispatches child", async () => {
-		const parentRunId = await h.makeRun("warren-parent");
-		await h.repos.runs.markRunning(parentRunId, NOW);
-		await h.repos.runs.finalize(parentRunId, "succeeded", NOW);
-		await h.repos.runs.setPrUrl(parentRunId, "https://github.com/x/y/pull/99");
-		const pr = await createPlanRunWithParent(parentRunId);
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: async () => ({ kind: "merged", mergedAt: NOW.toISOString() }),
-			spawn: h.spawnStub(() => "run_child"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-	});
-
-	test("parentRunId set, parent PR open → waiting_for_parent_merge", async () => {
-		const parentRunId = await h.makeRun("warren-parent");
-		await h.repos.runs.markRunning(parentRunId, NOW);
-		await h.repos.runs.finalize(parentRunId, "succeeded", NOW);
-		await h.repos.runs.setPrUrl(parentRunId, "https://github.com/x/y/pull/99");
-		const pr = await createPlanRunWithParent(parentRunId);
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: async () => ({ kind: "open" }),
-			spawn: h.spawnStub(() => "unused"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("waiting_for_parent_merge");
-		const reloaded = await h.repos.planRuns.require(pr.id);
-		expect(reloaded.state).toBe("running");
-	});
-
-	test("parentRunId set, parent PR closed unmerged → plan_failed", async () => {
-		const parentRunId = await h.makeRun("warren-parent");
-		await h.repos.runs.markRunning(parentRunId, NOW);
-		await h.repos.runs.finalize(parentRunId, "succeeded", NOW);
-		await h.repos.runs.setPrUrl(parentRunId, "https://github.com/x/y/pull/99");
-		const pr = await createPlanRunWithParent(parentRunId);
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: async () => ({ kind: "closed_unmerged" }),
-			spawn: h.spawnStub(() => "unused"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("plan_failed");
-		if (result.kind === "plan_failed") {
-			expect(result.reason).toBe("parent_pr_not_merged");
-		}
-		const reloaded = await h.repos.planRuns.require(pr.id);
-		expect(reloaded.state).toBe("failed");
-		expect(reloaded.failureReason).toBe("parent_pr_not_merged");
-	});
-
-	test("parentRunId set, parent has no PR + empty_push → trivial merge, gate passes", async () => {
-		const parentRunId = await h.makeRun("warren-parent");
-		await h.repos.runs.markRunning(parentRunId, NOW);
-		await h.repos.runs.finalize(parentRunId, "succeeded", NOW);
-		const seq = ((await h.repos.events.maxSeqForRun(parentRunId)) ?? 0) + 1;
-		await h.repos.events.append({
-			runId: parentRunId,
-			burrowEventSeq: seq,
-			ts: NOW.toISOString(),
-			kind: "reap.empty_push",
-			stream: "system",
-			payload: { commitsAhead: 0 },
-		});
-		const pr = await createPlanRunWithParent(parentRunId);
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "run_child"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-	});
-
-	test("parentRunId set, parent run deleted → gate passes (best-effort)", async () => {
-		const pr = await createPlanRunWithParent("run_nonexistent_zzz");
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "run_child"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("dispatched");
-	});
-
-	test("parentRunId set, parent still running → waiting_for_parent_merge", async () => {
-		const parentRunId = await h.makeRun("warren-parent");
-		await h.repos.runs.markRunning(parentRunId, NOW);
-		const pr = await createPlanRunWithParent(parentRunId);
-		const result = await advancePlanRun({
-			planRun: pr,
-			repos: h.repos,
-			showSeed: h.showSeedStub("open"),
-			checkPrMerged: neverPoll,
-			spawn: h.spawnStub(() => "unused"),
-			emit: h.emit,
-			now: () => NOW,
-		});
-		expect(result.kind).toBe("waiting_for_parent_merge");
 	});
 });
